pyproject.toml
[project]
name = "ai-engine"
version = "0.1.0"
description = ""
authors = [
    {name = "ApoorvZime",email = "apoorv@zime.ai"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "langgraph (>=1.0.5,<2.0.0)",
    "langchain (>=1.2.0,<2.0.0)",
    "langchain-core (>=1.2.5,<2.0.0)",
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "numpy (>=2.4.0,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "langchain-anthropic (>=1.3.0,<2.0.0)",
    "jsonschema (>=4.25.1,<5.0.0)",
    "langchain-openai (>=1.1.6,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
black = "^25.12.0"
isort = "^7.0.0"
ruff = "^0.14.10"

